adding data explorer, robots.txt, and loader
the data explorer has a data attributes (autocomplete) box on the left, meaning that the user
can search for one or more attributes (conditions, datatype, collection, medium, etc.) and
this can be used as is or in addition to a general query search. The search is done via an
ajax request so the page does not need to be updated. We also are adding a fader with loading
animation for queries that might take slightly longer. The robots.txt added will block all
agents at /, and a production version is available when we want to open up for robot
indexing.

Signed-off-by: vsoch <vsochat@stanford.edu>